Data.Aeson.Decoding.Tokens

Description

Token definitions.

Types

data Tokens k e

A well-formed JSON token stream.

Note: Lit exists to make Tokens have only 6 constructors. This may or may not have impact on performance.

Since: 2.1.2.0

Constructors

TkLit !Lit k

TkText !Text k

TkNumber !Number k

TkArrayOpen (TkArray k e)

TkRecordOpen (TkRecord k e)

TkErr e

data Lit

Literals. null, true, false.

Constructors

LitNull

LitTrue

LitFalse

data Number

Numbers

We preserve whether the number was integral, decimal or in scientific form.

Constructors

NumInteger !Integer
e.g. 123

NumDecimal !Scientific
e.g. 123.456

NumScientific !Scientific
e.g. 123e456, 123e-456 or 123.456E-967

data TkArray k e

Array tokens.

Constructors

TkItem (Tokens (TkArray k e) e)

TkArrayEnd k

TkArrayErr e

data TkRecord k e

Record tokens.

Constructors

TkPair !Key (Tokens (TkRecord k e) e)

TkRecordEnd k

TkRecordErr e