Data.Aeson.Encode

Description

Efficiently serialize a JSON value.

Most frequently, you'll probably want to encode straight to UTF-8 (the standard JSON encoding) using encode.

You can use the conversions to Builders when embedding JSON messages as parts of a protocol.

Documentation

encode :: ToJSON a => a -> ByteString

Efficiently serialize a JSON value as a lazy ByteString.

This is implemented in terms of the ToJSON class's toEncoding method.

Encoding to Builders

encodeToBuilder :: Value -> Builder

Encode a JSON value to a Data.ByteString Builder.

Use this function if you are encoding over the wire, or need to prepend or append further bytes to the encoded JSON value.

encodeToTextBuilder :: Value -> Builder

Encode a JSON Value to a Data.Text Builder, which can be embedded efficiently in a text-based protocol.

If you are going to immediately encode straight to a ByteString, it is more efficient to use encodeToBuilder instead.

Deprecated

fromValue :: Value -> Builder

Deprecated: Use encodeToTextBuilder instead