Data.Aeson.Generic

Description

JSON handling using Generics.

This is based on the Generic package originally written by Lennart Augustsson.

Decoding and encoding

decode :: Data a => ByteString -> Maybe a

Efficiently deserialize a JSON value from a lazy ByteString. If this fails due to incomplete or invalid input, Nothing is returned.

This function parses immediately, but defers conversion. See json for details.

decode' :: Data a => ByteString -> Maybe a

Efficiently deserialize a JSON value from a lazy ByteString. If this fails due to incomplete or invalid input, Nothing is returned.

This function parses and performs conversion immediately. See json' for details.

encode :: Data a => a -> ByteString

Efficiently serialize a JSON value as a lazy ByteString.

Lower-level conversion functions

fromJSON :: Data a => Value -> Result a

toJSON :: Data a => a -> Value