Data.Aeson.Key

Description

Strong type for JSON keys.

Since: 2.0.0.0

Documentation

data Key

fromString :: String -> Key

toString :: Key -> String

toText :: Key -> Text

fromText :: Text -> Key

coercionToText :: Maybe (Coercion Key Text)

'coercing r1 r2' will evaluate to r1 if Key is Coercible to Text, and to r2 otherwise.

Using coercing we can make more efficient implementations when Key is backed up by Text without exposing internals.

toShortText :: Key -> ShortText

Since: 2.0.2.0

fromShortText :: ShortText -> Key

Since: 2.0.2.0