Data.Aeson.KeyMap

Description

An abstract interface for maps from JSON keys to values.

Since: 2.0.0.0

Map Type

data KeyMap v

A map from JSON key type Key to v.

Query

null :: KeyMap v -> Bool

Is the map empty?

lookup :: Key -> KeyMap v -> Maybe v

Return the value to which the specified key is mapped, or Nothing if this map contains no mapping for the key.

(!?) :: KeyMap v -> Key -> Maybe v

Return the value to which the specified key is mapped, or Nothing if this map contains no mapping for the key.

This is a flipped version of lookup.

Since: 2.1.1.0

size :: KeyMap v -> Int

Return the number of key-value mappings in this map.

member :: Key -> KeyMap a -> Bool

Is the key a member of the map?

Construction

empty :: KeyMap v

Construct an empty map.

singleton :: Key -> v -> KeyMap v

Construct a map with a single element.

Insertion

insert :: Key -> v -> KeyMap v -> KeyMap v

Associate the specified value with the specified key in this map. If this map previously contained a mapping for the key, the old value is replaced.

insertWith :: (a -> a -> a) -> Key -> a -> KeyMap a -> KeyMap a

Insert with a function combining new and old values, taken in that order.

Since: 2.1.1.0

Deletion

delete :: Key -> KeyMap v -> KeyMap v

Remove the mapping for the specified key from this map if present.

Update

alterF :: Functor f => (Maybe v -> f (Maybe v)) -> Key -> KeyMap v -> f (KeyMap v)

alterF can be used to insert, delete, or update a value in a map.

Combine

difference :: KeyMap v -> KeyMap v' -> KeyMap v

Difference of two maps. Return elements of the first map not existing in the second.

union :: KeyMap v -> KeyMap v -> KeyMap v

unionWith :: (v -> v -> v) -> KeyMap v -> KeyMap v -> KeyMap v

The union with a combining function.

unionWithKey :: (Key -> v -> v -> v) -> KeyMap v -> KeyMap v -> KeyMap v

The union with a combining function.

intersection :: KeyMap a -> KeyMap b -> KeyMap a

The (left-biased) intersection of two maps (based on keys).

intersectionWith :: (a -> b -> c) -> KeyMap a -> KeyMap b -> KeyMap c

The intersection with a combining function.

intersectionWithKey :: (Key -> a -> b -> c) -> KeyMap a -> KeyMap b -> KeyMap c

The intersection with a combining function.

alignWith :: (These a b -> c) -> KeyMap a -> KeyMap b -> KeyMap c

Generalized union with combining function.

alignWithKey :: (Key -> These a b -> c) -> KeyMap a -> KeyMap b -> KeyMap c

Generalized union with combining function.

Lists

fromList :: [(Key, v)] -> KeyMap v

Construct a map with the supplied mappings. If the list contains duplicate mappings, the later mappings take precedence.

>>> fromList [("a", 'x'), ("a", 'y')]
fromList [("a",'y')]

fromListWith :: (v -> v -> v) -> [(Key, v)] -> KeyMap v

Construct a map from a list of elements. Uses the provided function, f, to merge duplicate entries with (f newVal oldVal).

toList :: KeyMap v -> [(Key, v)]

Return a list of this map's keys and elements.

The order is not stable. Use toAscList for stable ordering.

toAscList :: KeyMap v -> [(Key, v)]

Return a list of this map's elements in ascending order based of the textual key.

elems :: KeyMap v -> [v]

Return a list of this map' elements.

Since: 2.0.3.0

Maps

fromHashMap :: HashMap Key v -> KeyMap v

Convert a HashMap to a KeyMap.

toHashMap :: KeyMap v -> HashMap Key v

Convert a KeyMap to a HashMap.

fromHashMapText :: HashMap Text v -> KeyMap v

Convert a HashMap Textto a KeyMap.

toHashMapText :: KeyMap v -> HashMap Text v

Convert a KeyMap to a HashMap Text.

coercionToHashMap :: Maybe (Coercion (HashMap Key v) (KeyMap v))

fromMap :: Map Key v -> KeyMap v

Convert a Map to a KeyMap.

toMap :: KeyMap v -> Map Key v

Convert a KeyMap to a Map.

fromMapText :: Map Text v -> KeyMap v

Convert a Map Textto a KeyMap.

Since: 2.0.2.0

toMapText :: KeyMap v -> Map Text v

Convert a KeyMap to a Map Text.

Since: 2.0.2.0

coercionToMap :: Maybe (Coercion (Map Key v) (KeyMap v))

Traversal

Map

map :: (a -> b) -> KeyMap a -> KeyMap b

Map a function over all values in the map.

mapWithKey :: (Key -> a -> b) -> KeyMap a -> KeyMap b

Map a function over all values in the map.

Since: 2.1.0.0

mapKeyVal :: (Key -> Key) -> (v1 -> v2) -> KeyMap v1 -> KeyMap v2

Transform the keys and values of a KeyMap.

traverse :: Applicative f => (v1 -> f v2) -> KeyMap v1 -> f (KeyMap v2)

Perform an Applicative action for each key-value pair in a KeyMap and produce a KeyMap of all the results.

traverseWithKey :: Applicative f => (Key -> v1 -> f v2) -> KeyMap v1 -> f (KeyMap v2)

Perform an Applicative action for each key-value pair in a KeyMap and produce a KeyMap of all the results.

Folds

foldr :: (a -> b -> b) -> b -> KeyMap a -> b

foldr' :: (a -> b -> b) -> b -> KeyMap a -> b

foldl :: (b -> a -> b) -> b -> KeyMap a -> b

foldl' :: (b -> a -> b) -> b -> KeyMap a -> b

foldMapWithKey :: Monoid m => (Key -> a -> m) -> KeyMap a -> m

foldrWithKey :: (Key -> v -> a -> a) -> a -> KeyMap v -> a

Reduce this map by applying a binary operator to all elements, using the given starting value (typically the right-identity of the operator).

Conversions

keys :: KeyMap v -> [Key]

Return a list of this map's keys.

Filter

filter :: (v -> Bool) -> KeyMap v -> KeyMap v

Filter all keys/values that satisfy some predicate.

filterWithKey :: (Key -> v -> Bool) -> KeyMap v -> KeyMap v

Filter all keys/values that satisfy some predicate.

mapMaybe :: (a -> Maybe b) -> KeyMap a -> KeyMap b

Map values and collect the Just results.

mapMaybeWithKey :: (Key -> v -> Maybe u) -> KeyMap v -> KeyMap u

Map values and collect the Just results.

Key Type

data Key