Data.Aeson.Internal

Description

Internal types and functions.

Note: all declarations in this module are unstable, and prone to being changed at any time.

Documentation

data IResult a

The internal result of running a Parser.

Constructors

IError JSONPath String

ISuccess a

data JSONPathElement

Elements of a JSON path used to describe the location of an error.

type JSONPath = [JSONPathElement]

formatError :: JSONPath -> String -> String

Annotate an error message with a JSONPath error location.

ifromJSON :: FromJSON a => Value -> IResult a

Convert a value from JSON, failing if the types do not match.

iparse :: (a -> Parser b) -> a -> IResult b

Run a Parser.