Data.Aeson.Text

Description

Most frequently, you'll probably want to encode straight to UTF-8 (the standard JSON encoding) using encode.

You can use the conversions to Builders when embedding JSON messages as parts of a protocol.

Documentation

encodeToLazyText :: ToJSON a => a -> Text

Encode a JSON Value to a Data.Text.Lazy

Note: uses toEncoding

encodeToTextBuilder :: ToJSON a => a -> Builder

Encode a JSON Value to a Data.Text Builder, which can be embedded efficiently in a text-based protocol.

If you are going to immediately encode straight to a ByteString, it is more efficient to use encodeToBuilder instead.

Note: Uses toJSON